% Generated by roxygen2 (4.0.2): do not edit by hand
\name{adaptiveSimpsons}
\alias{adaptiveSimpsons}
\title{Adaptive Simpson's Rule}
\usage{
adaptiveSimpsons(fn, a, b, eps, subdivisions = 1024)
}
\arguments{
\item{fn}{The function to integrate}

\item{a}{The lower bound of integration}

\item{b}{The Upper bound of integration}

\item{eps}{The tolerance for stopping the integration routine}

\item{subdivisions}{The maximum number of simpson's rule estimates
to compute on the interval [a, b]}
}
\description{
This algorithm was derived from a publication by the
MathWorks team that produces Matlab and class notes from Simon Foucart's
Introduction to Numerical Mathematics course
}

