% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calcConf.discrete}
\alias{calcConf.discrete}
\title{Calculate the Bayesian confidence using the discrete Binomial cdf}
\usage{
calcConf.discrete(N, n1, n2, r, beta, lambda, maxN)
}
\arguments{
\item{N}{Population size}

\item{n1}{Number of judgmental samples}

\item{n2}{Number of random samples}

\item{r}{A high risk judgment sample location is expected to be \code{r}
times more likely to be contaminated than a low risk random sample location}

\item{beta}{The second shape parameter of the prior distribution (which is a
Beta distribution).}

\item{lambda}{The fraction of percentage clean in (0,1].}

\item{maxN}{A large integer indicating the transition from finite population
to infinite population formulas.}
}
\value{
The Bayesian confidence
}
\description{
Calculate the Bayesian confidence using the discrete Binomial cdf, i.e.
requiring the number of unacceptable items to be a whole number
}
\examples{
calcConf.discrete(1000, 10, 80, 2, 23, 0.99, 10^6)
calcConf.discrete(999, 10, 80, 2, 23, 0.99, 10^6)
}
\author{
Landon Sego
}
\seealso{
\code{\link{calcConf.raw}}, \code{\link{calcConf}}
}
\keyword{misc}

