% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calcConf.fnr}
\alias{calcConf.fnr}
\title{Calculate the CJR confidence for non-zero false negative rates}
\usage{
calcConf.fnr(n2, ..., arg.check = TRUE, nProcesses = detectCores() - 1,
  verbose = FALSE)
}
\arguments{
\item{n2}{The number of random samples}

\item{arg.check}{\code{=TRUE} checks the arguments supplied in \code{...}
using \code{\link{cjrParms}}.  I would not recommend setting this to
\code{FALSE}.}

\item{nProcesses}{Integer value of the number of cores to use in calculating the confidence in
parallel.  Defaults to one less than the total number of cores on the machine.}

\item{verbose}{\code{=TRUE} prints details from the calculation of the confidence}

\item{\dots}{Arguments to \code{\link{cjrParms}}}
}
\value{
The value of the Bayesian confidence
}
\description{
Calculates the Bayesian posterior probability of the number of contaminated
cells being less than a threshold, when the false negative rate is greater than zero
}
\details{
Using the posterior predictive distribution, calculates the probability that
the number of contaminated cells is less than a determined threshold, given
that none of the judgmental nor random samples detected the presence of
contamination, and accounting for the false negative error rate.

If \code{nProcesses = 1} an
unparallized version that uses a slightly different algorithm is used,
and this can serve as a check for the parallelized version.
}
\examples{
calcConf.fnr(239, fnr = 0, n1 = 10, N = 2500, pct.clean = 0.99, r = 2, prior.prob = 0.05)
calcConf.fnr(239, fnr = 0.1, n1 = 10, N = 2500, pct.clean = 0.99, r = 2, prior.prob = 0.05)
calcConf.fnr(239, fnr = 0.4, n1 = 10, N = 2500, pct.clean = 0.99, r = 2, prior.prob = 0.05)
calcConf.fnr(239, fnr = 0.7, n1 = 10, N = 2500, pct.clean = 0.99, r = 2, prior.prob = 0.05, nProcesses = 1)
calcConf.fnr(239, fnr = 0.7, n1 = 10, N = 2500, pct.clean = 0.99, r = 2, prior.prob = 0.05, nProcesses = 2)
}
\author{
Landon Sego
}
\seealso{
\code{\link{calcConf}}
}
\keyword{misc}

