% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calcConf.raw}
\alias{calcConf.raw}
\title{Calculate the Bayesian confidence and the sign of the first derivative}
\usage{
calcConf.raw(N, n1, n2, r, beta, lambda, maxN, output = c("confidence",
  "first.sign"))
}
\arguments{
\item{N}{Population size}

\item{n1}{Number of judgmental samples}

\item{n2}{Number of random samples}

\item{r}{A high risk judgment sample location is expected to be \code{r}
times more likely to be contaminated than a low risk random sample location}

\item{beta}{The second shape parameter of the prior distribution (which is a
Beta distribution).}

\item{lambda}{The fraction of percentage clean in (0,1].}

\item{maxN}{A large integer indicating the transition from finite population
to infinite population formulas.}

\item{output}{Character vector indicating which quanity to calculate: the
confidence, the first derivative, the sign of the first derivative, or any
combination thereof.}
}
\value{
List with the following elements: \item{confidence}{The Bayesian
confidence} \item{first.sign}{The sign of the partial derivative of the
confidence function wrt N evaluated at \code{N}}
}
\description{
Calculate the Bayesian confidence and the sign of the first derivative of
the confidence function wrt N at the value N.
}
\examples{
calcConf.raw(10^4, 10, 80, 2, 23, 0.98, 10^6, output="confidence")
calcConf.raw(10^4, 10, 95, 2, 23, 0.98, 10^6)
}
\author{
Landon Sego
}
\seealso{
\code{\link{calcConf}}, \code{\link{calcConf.discrete}}
}
\keyword{misc}

