% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{cjr-package}
\alias{cjr}
\alias{cjr-package}
\title{Routines for combined judgmental random sampling (CJR)}
\description{
Collection of functions for calculating Bayesian probabilities and sample
sizes for the CJR methodology
}
\details{
\tabular{ll}{
Package: \tab cjr\cr
Type: \tab Package\cr
Version: \tab 0.2.1\cr
Date: \tab 2015-01-20\cr
License: \tab PNNL business sensitive\cr}
}
\author{
Landon Sego

Maintainer:  Landon Sego <Landon.Sego@pnnl.gov>
}
\keyword{package}

