% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cjr.conf}
\alias{cjr.conf}
\title{Compute the confidence for the original CJR model}
\usage{
cjr.conf(lambda, m, nh, rho, Bh, N)
}
\arguments{
\item{lambda}{the upper tolerance limit for the percentage of acceptable,
low risk items (technically 1 - lambda)}

\item{m}{the number of low risk, random samples}

\item{nh}{the number of high-risk, judgmental samples}

\item{rho}{the expectation relation between the low-risk samples and
high risk samples such that one is rho times more likely to observe
an unacceptable \code{nh} than an unaccpetable \code{m}}

\item{Bh}{the hyperparameter for the beta prior distribution}

\item{N}{the total number of items in the population}
}
\description{
Compute the confidence for the original CJR model
}
\examples{
# I want no more than 1\% of all items in the population to be unacceptable
# (same as saying 99\% of items acceptable)
lambda <- 0.99

# Experts have decided that there are 30 high-risk items, all of which must
# be sampled and verified clean
nh <- 30

# There are 5000 total grid cells in the population
N <- 5000

# Experts determine one is 2x as likely to observe an unacceptable
# item in the high-risk sample relative to the low-risk sample
rho <- 2

# Experts believe, a priori, that there is a 1/100 probability of
# observing an unacceptable item in the high risk population. To estimate
# the hyperparameter: (1 - 1/100)/(1/100) = 99
Bh <- 99

# Inverting the problem shows that n = 38 random samples are required to
# reach a target confidence of 0.95
n <- 38

Conf <- cjr.conf(0.99, 38, 30, 2, 99, 5000)
}
\author{
Alex Venzin
}

