% Generated by roxygen2 (4.0.2): do not edit by hand
\name{decreasingConf}
\alias{decreasingConf}
\title{Validates that the confidence function is deceasing for "all" N}
\usage{
decreasingConf(lambda, n1, n2, r, beta, maxN)
}
\arguments{
\item{lambda}{The percentage clean.}

\item{n1}{The number of judgmental samples}

\item{n2}{The number of random samples}

\item{r}{Ratio of prior expecation between judgmental and random samples}

\item{beta}{Second shape pararments of the beta prior of the judgmental
samples}

\item{maxN}{A large integer indicating the transition from finite population
to infinite population formulas.}
}
\value{
A logical (boolean) indicating whether the confidence function is
decreasing for "all" \code{N} in \code{[Nmin, maxN]} where \code{Nmin} is
the smallest \code{N} can be given \code{n1}, \code{n2}, and \code{lambda}.

This sequence of \code{N} is linear logarithmic: it is more dense for
smaller \code{N} and less dense for larger \code{N}, i.e., the natural
logarithm of the sequence has values that are evenly spaced.
}
\description{
A helper function for \code{\link{is.viable.lambda}} which determines
whether the confidence function is deceasing for a sequence of \code{N}.
}
\examples{
decreasingConf(0.99, 10, 30, 2, 200, 10^6)
decreasingConf(0.995, 10, 30, 2, 200, 10^6)
}
\author{
Landon Sego
}
\seealso{
Called by \code{\link{is.viable.lambda}}.
}
\keyword{misc}

