% Generated by roxygen2 (4.0.2): do not edit by hand
\name{find.n2.initial}
\alias{find.n2.initial}
\title{Finds the number of random samples needed to satisfy CJR input parameters
using the initial CJR model}
\usage{
find.n2.initial(conf, ..., method = c("cont.binom", "floor"),
  int.method = c("quadrature", "simpson"), simpson.length = 3001,
  round.up = FALSE, arg.check = TRUE)
}
\arguments{
\item{conf}{A number in (0,1) giving the desired Bayesian confidence}

\item{method}{Only applies to the initial CJR model.  Method used to
calculate probabilities and sample sizes, see Details below.}

\item{int.method}{Only applies to the initial CJR model.  Indicates whether
numerical integrations are performed using Gaussian quadrature or Simpson's
method.  See Details below.}

\item{simpson.length}{Only applies to the initial CJR model.  An odd integer
indicating the number points at which the integrand function is evaluated
within the interval of integration when using Simpson's method. See Details
below.}

\item{round.up}{\code{=TRUE} rounds the sample size up to the nearest whole
integer}

\item{arg.check}{\code{=TRUE} checks the arguments supplied in \code{...}
using \code{\link{cjrParms}}}

\item{\dots}{Arguments to \code{\link{cjrParms}}}
}
\value{
A list with the following components: \item{n2}{the number of random
samples} \item{conf.achieved}{The actual Bayesian confidence achieved by
\code{n2}}
}
\description{
Finds the number of random samples needed to satisfy CJR input parameters,
using the initial CJR model of theta and theta/r.  Note, however, that the
CJR implementation in VSP is based on \code{\link{find.n2}}.
}
\details{
When calling \code{find.n2.initial} directly, \code{arg.check} should always
be \code{TRUE}.
}
\examples{
# These produce the same result, since they use the same prior
# distribution, because beta = (1 - prior.prob) / prior.prob
find.n2.initial(0.95, n1=10, N=25000, pct.clean=0.99, r=2, prior.prob=0.05)
find.n2.initial(0.95, n1=10, N=25000, pct.clean=0.99, r=2, beta=19)
}
\author{
Landon Sego
}
\seealso{
\code{find.n2} calls: \code{\link{cjrParms}} and
\code{\link{probRemain}}
}
\keyword{misc}

