% Generated by roxygen2 (4.0.2): do not edit by hand
\name{find.n2}
\alias{find.n2}
\title{Finds the number of random samples needed to satisfy CJR input parameters}
\usage{
find.n2(conf, ..., maxN = 10^6, round.up = FALSE,
  search.viable.lambda = TRUE, decimal.precision = 5, arg.check = TRUE,
  show.viable.plot = FALSE, discrete = FALSE)
}
\arguments{
\item{conf}{A number in (0,1) giving the desired Bayesian confidence}

\item{maxN}{A large integer indicating the transition from finite population
to infinite population formulas.}

\item{round.up}{\code{=TRUE} rounds the sample size up to the nearest whole
integer.}

\item{search.viable.lambda}{\code{=TRUE} Searches for a large enough value
of the percentage clean so that the confidence function is always a
non-increasing function of N.}

\item{decimal.precision}{The level of precision for the viable lambda,
passed to \code{\link{find.viable.lambda}}}

\item{arg.check}{\code{=TRUE} checks the arguments supplied in \code{...}
using \code{\link{cjrParms}}.  I would not recommend setting this to
\code{FALSE}.}

\item{show.viable.plot}{\code{=TRUE} makes a plot which illustrates the
identification of the viable percentage clean value.}

\item{discrete}{\code{=TRUE} uses \code{\link{calcConf.discrete}} to
calculate the confidence when determining the sample size.}

\item{\dots}{Arguments to \code{\link{cjrParms}}}
}
\value{
A list with the following elements: \item{n2}{The required number of
random samples} \item{conf.acheived}{The confidence acheived by the
\code{n2}} \item{prior.pct.clean}{The expected value of the percentage clean
prior to collecting data, which depends the choices for \code{N}, \code{n1},
\code{prior.prob} (or \code{beta}), and \code{r}.} \item{use.pct.clean}{The
percentage clean used to calculate \code{n2}}
\item{recommended.pct.clean}{The percentage clean that will ensure a
non-increasing confidence function for the given parameter inputs.}
\item{is.viable.request}{A logical indicating whether the requested
\code{pct.clean} is viable (i.e. results in a confidence function that is
non-decreasing in \code{N}).}
}
\description{
Finds the number of random samples needed to satisfy CJR input parameters
}
\examples{
find.n2(0.95, n1=10, N=10^4, pct.clean=0.99, prior.prob=0.05, r=1.5, show.viable.plot = TRUE)
find.n2(0.95, n1=10, N=10^4, pct.clean=0.99, prior.prob=0.01, r=3, show.viable.plot = TRUE)
}
\author{
Landon Sego
}
\seealso{
\code{\link{find.n2.viable}}
}
\keyword{misc}

