% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fnrConf.hgeo}
\alias{fnrConf.hgeo}
\title{Calculate the confidence that at most t samples are unacceptable in
the unsampled population}
\usage{
fnrConf.hgeo(n2, ..., arg.check = TRUE, subdivisions = 1e+05 + 1,
  method = c("C", "R"))
}
\arguments{
\item{n2}{The number of random samples}

\item{subdivisions}{The number of points to sample in the simpson's rule
approximation}

\item{method}{One of 'C' or 'R'. The default option is to export the
computation to C (much faster).}

\item{n1}{The number of judgmental samples}

\item{fnr}{The false negative rate associated with the measurements}

\item{prior.prob}{The prior probability of observing an unacceptable
item in the judgmental population}

\item{N}{The total number of items in the population}

\item{r}{The likelihood multiplier than an item sampled from the
judgmental subpopulation is going to be unacceptable relative
to the random subpopulation}
}
\description{
Calculate the confidence that at most t samples are unacceptable in
the unsampled population
}
\examples{
# There are 10 judgmental items
nh <- 10

# The prior probability of observing an unacceptable item
# in nh items is 1\%
prior.prob <- .01

# I will take 50 random samples
nl <- 50

# The probability my measurement device/algorithm produces
# a false positive is, on average, 0.1
fnr <- 0.1

# I believe that the nh items are twice as likely as the nl
# items to be unacceptable
r <- 2

# There are 500 total items in the population
N <- 500

# I will accept at most 5 unacceptable items remaining
# in the population after sampling nh + nl items out of N
thresh <- 5

# The confidence is...
conf <- fnrConf.hgeo(n1 = nh, n2 = nl, fnr = fnr, N = N,
                     r = r, thresh = thresh, prior.prob = prior.prob)
}

