% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gCalcConf}
\alias{gCalcConf}
\title{Simulate the confidence of the generalized CJR}
\usage{
gCalcConf(n1, nh, n2, N, beta, rho, p, lambda, c.prime = 0.95, nreps = 10^4)
}
\arguments{
\item{n1}{Number of judgmental samples taken (perceived number of high risk
cells)}

\item{nh}{True number of high-risk cells}

\item{n2}{Number of random samples taken from unsampled areas, which are
presumed by the investigator to be low-risk cells.}

\item{N}{Population size}

\item{beta}{The true value of \deqn{\frac{1-E(\theta_h)}{E(\theta_h)}}{(1 -
E(\theta_h)) / E(\theta_h)}}

\item{rho}{The true value of
\deqn{\frac{E(\theta_h)}{E(\theta_l)}}{E(\theta_h) / E(\theta_l)}}

\item{p}{The probability that a judgmental sample is placed in a high-risk
cell.}

\item{lambda}{The nominal fraction of the population that we desire to show
is acceptable}

\item{c.prime}{The desired confidence level}

\item{nreps}{The number of Monte Carlo replicates}
}
\value{
A list with the following components: \item{conf}{The simulated
confidence} \item{se}{The standard error of the confidence estimate}
\item{nreps}{The number of Monte Carlo replicates} \item{z}{The z-score
comparing the simulated \code{conf} and the desired \code{c.prime}}
}
\description{
Simulate the true confidence using a generalization of the CJR method.
Useful for sensitivity studies regarding misspecification of the prior
parameters, the number and location of judgmental samples.
}
\examples{
gCalcConf(25, 18, 75, 10^4, 90, 2, 0.80, 0.99)
}
\author{
Landon Sego
}
\seealso{
\code{\link{calcConf}}
}
\keyword{misc}

