% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.viable.lambda}
\alias{is.viable.lambda}
\title{Determine whether lambda produces a non-increasing confidence function}
\usage{
is.viable.lambda(target.lambda, n1, n2, r, beta, maxN, adjust = 0,
  decimal.precision = 5, verbose = FALSE)
}
\arguments{
\item{target.lambda}{The value of \code{lambda} to be tested.}

\item{n1}{Number of judgmental samples}

\item{n2}{Number of random samples}

\item{r}{Ratio of prior expecation between judgmental and random samples}

\item{beta}{Second shape parameter of the beta prior of the judgmental
samples}

\item{maxN}{A large integer indicating the transition from finite population
to infinite population formulas. This argument is now deprecated (not used)
but I've included it for backwards compatability with earlier versions.}

\item{adjust}{This is a small adjustment on the first condition so as to not
disqualify a viable lambda due to machine error.  The only time this value
should be greater than 0 is during validation excercises--e.g.
\code{\link{minimum.viable.lambda}} calls \code{is.viable.lambda} using
\code{adjust = 1e-14}.  This argument is now deprecated (not used) but I've
included it for backwards compatability with earlier versions.}

\item{decimal.precision}{The level of precision that will be used to compare
the \code{target.lambda} and the calculated viable lambda.}

\item{verbose}{Prints details. This argument is now deprecated (not used)
but I've included it for backwards compatability with earlier versions.}
}
\value{
A list with the following elements \item{viable}{A logical (Boolean)
indicating whether \code{target.lambda} is viable (i.e. satisfies the
theorem)} \item{msg}{If \code{viable=FALSE}, a list containing the input
values of the function that might be used for diagnosis.}
}
\description{
Determine whether \code{lambda} produces a non-increasing confidence
function
}
\details{
Determines whether the value of \code{target.lambda} is viable by applying
the theorem.
}
\examples{
is.viable.lambda(0.90, 10, 30, 3, 300, 10^6)
is.viable.lambda(0.99, 10, 30, 2, 5, 10^6, verbose=TRUE)
}
\author{
Landon Sego
}
\references{
PNNL-19315
}
\seealso{
\code{\link{find.viable.lambda}}
}
\keyword{misc}

