% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{journals}
\alias{journals}
\title{CEFAGE-UE journals' ranking}
\format{A data frame with 913 rows (journals) and 6 variables:
\describe{
  \item{journal}{Journal name.}
  \item{ISSN}{Journal's
  \href{https://en.wikipedia.org/wiki/International_Standard_Serial_Number}{International
  Standard Serial Number (ISSN)}.}
  \item{eISSN}{Journal's
  \href{https://en.wikipedia.org/wiki/International_Standard_Serial_Number#Electronic_ISSN}{electronic
  International Standard Serial Number (eISSN)}.}
  \item{publishing_press_abbr}{Journal's publishing press abbreviation. See
  dataset \code{\link[cjr]{publishing_presses}} for the full name of each
  publishing press.}
  \item{scientific_field_abbr}{Journal's scientific field abbreviation. See
  dataset \code{\link[cjr]{scientific_fields}} for the full name of each
  scientific discipline(s).}
  \item{rate}{Journal's rating according to CEFAGE. See dataset
  \code{\link[cjr]{rates}} for the meaning of each rate.}
}}
\source{
\url{http://www.cefage.uevora.pt/en/content/download/5988/71181/version/1/file/RankingCEFAGE-4ed.pdf}
}
\usage{
journals
}
\description{
A dataset containing the ranking of scientific journals, 4th Edition
(2016-2019), authored and provided by the
\href{http://www.cefage.uevora.pt}{Center for Advanced Studies in Management
and Economics (CEFAGE)}, University of Évora, Portugal.
}
\details{
This dataset has been obtained by parsing the original source (PDF).
}
\keyword{datasets}
