% Generated by roxygen2 (4.0.2): do not edit by hand
\name{min.lamb.cjr}
\alias{min.lamb.cjr}
\title{Compute minimum viable lambda}
\usage{
\method{min}{lamb.cjr}(lambda, n, nh, xh, rho, Bh, pl, N)
}
\arguments{
\item{lambda}{the upper tolerance limit, i.e, "I desire that no more than (1
- lambda)*N of the low-risk, unsampled items in the population are
unacceptable."}

\item{n}{the number of low-risk, random samples}

\item{nh}{the number of high-risk, judgmental samples}

\item{xh}{the estimated number of unacceptable, high-risk items present in
the population}

\item{rho}{the expectation relation that governs the relationship between
sampled, judgmental items and low-risk, random items in the population}

\item{Bh}{the hyperparameter for theta_h}

\item{pl}{the estimated probability that the measurement device or method
incorrectly determines that an unacceptable low-risk item is acceptable}

\item{N}{the total number of items in the population (or sample space).}
}
\value{
The minimum viable lambda such that the confidence function is
strictly non-decreasing. Note that if the supplied lambda is already
sufficient, then lmin = lambda. In any case, lmin >= lambda.
}
\description{
Computes the minimum lambda such that the posterior predictive distribution
is a strictly non-decreasing function of the supplied parameters.
}
\author{
Alex Venzin
}

