% Generated by roxygen2 (4.0.2): do not edit by hand
\name{minimum.viable.lambda}
\alias{minimum.viable.lambda}
\title{Determine whether "all" lambda greater than a known viable lambda are also
viable}
\usage{
minimum.viable.lambda(conf, n1, n2, r, beta, viable.lambda, maxN,
  max.seq.length = 50, verbose = FALSE)
}
\arguments{
\item{conf}{The value of the Bayesian confidence}

\item{n1}{The number of judgmental samples}

\item{n2}{The number of random samples}

\item{r}{Ratio of prior expecation between judgmental and random samples}

\item{beta}{Second shape pararments of the beta prior of the judgmental
samples}

\item{viable.lambda}{A known viable \code{lambda}, likely obtained from
\code{\link{find.viable.lambda}}, or verifed to be viable using
\code{\link{is.viable.lambda}}.}

\item{maxN}{A large integer indicating the transition from finite population
to infinite population formulas.}

\item{max.seq.length}{The maximum number of \code{lambda} values in
[\code{viable.lambda}, 1] to test for viability}

\item{verbose}{\code{=TRUE} prints intermediate results}
}
\value{
A list with the following components \item{all.viable}{A logical
indicating whether all the \code{lambda} in [\code{viable.lambda}, 1] were
viable} \item{num.not.viable}{The count of the number of times where the
\code{lambda} was not viable} \item{seq.length}{The length of the sequence
of \code{lambds} that was tested.} \item{pct.viable}{The percentage of the
tested \code{lambda} values that were viable} \item{results}{A data frame
showing the detailed results for the testing of each \code{lambda}.}
}
\description{
A validation function to determine whether "all" \code{lambda} greater than
a known viable \code{lambda} are also viable
}
\details{
Checks the viability of a sequence of \code{lambda} values in the interval
[\code{viable.lambda}, 1].
}
\examples{
v.lambda <- find.viable.lambda(0.8, 0.95, 10, 72, 2, 392, 10^6)
pvar(v.lambda)
minimum.viable.lambda(0.95, 10, 72, 2, 392, v.lambda, 10^6)
}
\author{
Landon Sego
}
\seealso{
\code{minimum.viable.lambda} is called by
\code{\link{n2.vs.viable.lambda}}.
}
\keyword{misc}

