% Generated by roxygen2 (4.0.2): do not edit by hand
\name{paramSplit}
\alias{paramSplit}
\title{Split the domain of integration for adaptive Simpson's Rule}
\usage{
paramSplit(params)
}
\arguments{
\item{params}{A set of closed intervals. Each interval will be split in two
such that [a, b] = {[a, c], [c, b]} where c = midpoint(a, b)}
}
\description{
Split the domain of integration for adaptive Simpson's Rule
}
\note{
It is assumed that a <= b for each subinterval.
If this is not true, calculate the integral multiplied by -1.
}

