% Generated by roxygen2 (4.0.2): do not edit by hand
\name{probRemain}
\alias{probRemain}
\title{Calculates the Bayesian posterior probability of the number of contaminated
cells being less than a threshold, using initial CJR model}
\usage{
probRemain(n2, ..., method = c("cont.binom", "floor"),
  int.method = c("quadrature", "simpson"), simpson.length = 3001,
  plot.integrands = FALSE, simpson.check = FALSE, arg.check = TRUE)
}
\arguments{
\item{n2}{The number of random samples}

\item{method}{Only applies to the initial CJR model.  Method used to
calculate probabilities and sample sizes, see Details below.}

\item{int.method}{Only applies to the initial CJR model.  Indicates whether
numerical integrations are performed using Gaussian quadrature or Simpson's
method.  See Details below.}

\item{simpson.length}{Only applies to the initial CJR model.  An odd integer
indicating the number points at which the integrand function is evaluated
within the interval of integration when using Simpson's method. See Details
below.}

\item{plot.integrands}{\code{=TRUE} produces a plot (using the X11 device)
of the integrands of the numerator and denominator of the Bayesian
confidence.  \code{="filename.pdf"} produces the same plot but writes it to
"filename.pdf".}

\item{simpson.check}{\code{=TRUE} checks the integration results from the
quadrature using Simpsons' method.  If the results differ, a detailed
message is printed.  If \code{int.method='simpson'} is specified in
\code{...}, then no check is performed.}

\item{arg.check}{\code{=TRUE} checks the arguments supplied in \code{...}
using \code{\link{cjrParms}}}

\item{\dots}{Arguments to \code{\link{cjrParms}}}
}
\value{
The probability that the sampling area is \code{pct.clean} x 100\%
clean, given that none of the judgmental nor random samples detected the
presence of contamination.
}
\description{
Using the posterior predictive distribution, calculates the probability that
the number of contaminated cells is less than a determined threshold, given
that none of the judgmental nor random samples detected the presence of
contamination.  This follows the initial CJR model which uses theta and
theta/r.  Note, however, that the CJR implementation in VSP is based on
\code{\link{calcConf}}.
}
\details{
When calling \code{probRemain} directly, \code{arg.check} should always be
\code{TRUE}.  It is set to false when it is called by \code{\link{find.n2}}.

For the initial CJR model where P(unacceptable judgmental sample) = theta
and P(unacceptable random sample) = theta/r, the functions
\code{\link{find.n2.initial}} and \code{\link{probRemain}} can be used.  The
arguments \code{method}, \code{int.method}, and \code{simpson.length} are
only applicable to those two functions.

When chosing \code{method = "floor"}, the number of tolerable contaminated
cells is chosen using \code{floor((1 - pct.clean) * N)}, and the standard,
discrete binomial distribution is used to calculate probabilities.  This
results in confidence statements like "95\% confident that at least 99\% of
the grid cells do not contain detectable contamination."

If \code{method = "cont.binom"}, the number of tolerable contaminated cells
is simply \code{(1 - pct.clean) * N}, and a continuous version of the
binomial distribution is used to calculate probabilities.  Results in
confidence statements like "95\% confident that 99\% of the decision area
does not contain detectable contamination."
}
\examples{
# These produce the same result, since they use the same prior
# distribution, because beta = (1 - prior.prob) / prior.prob
probRemain(239, n1=10, N=25000, pct.clean=0.99, r=2, prior.prob=0.05)
probRemain(239, n1=10, N=25000, pct.clean=0.99, r=2, beta=19)
}
\author{
Landon Sego
}
\seealso{
\code{probRemain} called by: \code{\link{find.n2}}, Calls:
\code{\link{cjrParms}}
}
\keyword{misc}

