% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{publishing_presses}
\alias{publishing_presses}
\title{Publishing presses}
\format{A data frame with 34 rows (publishing presses) and 2 variables:
\describe{
  \item{publishing_press_abbr}{Abbreviation of publishing press.}
  \item{publishing_press}{Publishing press.}
}}
\source{
\url{http://www.cefage.uevora.pt/en/content/download/5988/71181/version/1/file/RankingCEFAGE-4ed.pdf}
}
\usage{
publishing_presses
}
\description{
A dataset containing the different publishing presses and its abbreviations.
}
\details{
This dataset has been obtained by parsing the original source (PDF).
}
\keyword{datasets}
