% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rates}
\alias{rates}
\title{CEFAGE-UE journals' ratings}
\format{A data frame with 6 rows (rates) and 2 variables:
\describe{
  \item{rate}{Rate value.}
  \item{description}{Rate description.}
}}
\source{
\url{http://www.cefage.uevora.pt/en/content/download/5988/71181/version/1/file/RankingCEFAGE-4ed.pdf}
}
\usage{
rates
}
\description{
A dataset containing the different ratings used to classify each journal.
This rating system is by the \href{http://www.cefage.uevora.pt}{Center for
Advanced Studies in Management and Economics (CEFAGE)}, University of Évora,
Portugal.
}
\details{
This dataset has been obtained by parsing the original source (PDF).
}
\keyword{datasets}
