% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sampleSize.fnr}
\alias{sampleSize.fnr}
\title{Compute the minimum sample size to reach a specified target confidence using either the
exact method or the asymptotic approximation.}
\usage{
sampleSize.fnr(N, nh, fnr, rho, thresh, prior.prob, targetConf,
  method = c("discrete", "continuous"), nProcesses = 1)
}
\arguments{
\item{N}{The total number of items in the population}

\item{nh}{The number of judgmental samples}

\item{fnr}{The false negative probability}

\item{rho}{The expectation relation multiplier}

\item{thresh}{The number of items that one is willing to accept following sampling that
are unacceptable}

\item{prior.prob}{The prior probability of observing an unacceptable item in the judgmental
subpopulation}

\item{targetConf}{The target confidence you are trying to achieve}

\item{method}{Can be one of 'exact' or 'approximation'. If \code{(1-pct.clean)*N}
is large, then the approximation method is significantly faster, but less accurate}

\item{nProcesses}{The number of parallel processes required to calculate the exact confidence}
}
\description{
Compute the minimum sample size to reach a specified target confidence using either the
exact method or the asymptotic approximation.
}
\author{
Alex Venzin
}

