% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scientific_fields}
\alias{scientific_fields}
\title{Scientific fields}
\format{A data frame with 32 rows (scientific fields) and 2 variables:
\describe{
  \item{scientific_field_abbr}{Abbreviation of scientific field(s).}
  \item{scientific_field}{Scientific field(s).}
}}
\source{
\url{http://www.cefage.uevora.pt/en/content/download/5988/71181/version/1/file/RankingCEFAGE-4ed.pdf}
}
\usage{
scientific_fields
}
\description{
A dataset containing the different scientific fields and its abbreviations.
}
\details{
This dataset has been obtained by parsing the original source (PDF).
}
\keyword{datasets}
