% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simpsonsRule}
\alias{simpsonsRule}
\title{The 3 point Simpson's Rule}
\usage{
simpsonsRule(fn, a, b, h)
}
\arguments{
\item{fn}{The function to integrate}

\item{a}{The lower bound of integration}

\item{b}{The upper bound of integration}

\item{h}{(b-a)/2}
}
\description{
The 3 point Simpson's Rule
}

