% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sRecurse}
\alias{sRecurse}
\title{Recursively apply simpson's rule to a set of closed intervals}
\usage{
sRecurse(fn, params, h)
}
\arguments{
\item{fn}{The function to integrate}

\item{params}{The set of intervals on which to apply
the 3 point Simpson's Rule}

\item{h}{(x_i - x_j)/2 where x_i > x_j and [x_j, x_i]
is a subinterval contained in params}
}
\description{
Recursively apply simpson's rule to a set of closed intervals
}

